+++ apps/netconf/netconf_main.c
@@ -280,6 +280,8 @@ netconf_terminate(clicon_handle h)
     clicon_rpc_close_session(h);
     if ((yspec = clicon_dbspec_yang(h)) != NULL)
 	yspec_free(yspec);
+    if ((yspec = clicon_config_yang(h)) != NULL)
+	yspec_free(yspec);
     if ((x = clicon_conf_xml(h)) != NULL)
 	xml_free(x);
     event_exit();
@@ -390,7 +392,7 @@ main(int    argc,
     /* Find and read configfile */
     if (clicon_options_main(h, yspecfg) < 0)
 	return -1;
-
+    clicon_config_yang_set(h, yspecfg);
     /* Now rest of options */
     optind = 1;
     opterr = 0;