+++ apps/restconf/restconf_main.c
@@ -454,6 +454,8 @@ restconf_terminate(clicon_handle h)
     clicon_rpc_close_session(h);
     if ((yspec = clicon_dbspec_yang(h)) != NULL)
 	yspec_free(yspec);
+    if ((yspec = clicon_config_yang(h)) != NULL)
+	yspec_free(yspec);
     if ((x = clicon_conf_xml(h)) != NULL)
 	xml_free(x);
     clicon_handle_exit(h);
@@ -583,6 +585,7 @@ main(int    argc,
     /* Find and read configfile */
     if (clicon_options_main(h, yspecfg) < 0)
 	goto done;
+    clicon_config_yang_set(h, yspecfg);
     stream_path = clicon_option_str(h, "CLICON_STREAM_PATH");
     /* Now rest of options, some overwrite option file */
     optind = 1;