+++ lib/clixon/clixon_options.h
@@ -165,6 +165,11 @@ int clicon_quiet_mode_set(clicon_handle h, int val);
 yang_spec * clicon_dbspec_yang(clicon_handle h);
 int clicon_dbspec_yang_set(clicon_handle h, struct yang_spec *ys);
 
+#if 1 /* Temporary function until "Top-level Yang symbol cannot be called "config"" is fixed */
+yang_spec * clicon_config_yang(clicon_handle h);
+int clicon_config_yang_set(clicon_handle h, struct yang_spec *ys);
+#endif
+
 cxobj *clicon_conf_xml(clicon_handle h);
 int clicon_conf_xml_set(clicon_handle h, cxobj *x);